reports_mjt/requirements.txt
Django==1.11.11
PyMySQL==0.9.3
djangorestframework
pandas==0.20.3
django_redis
redis==3.2.1
xlwt==1.3.0
celery==4.3.0
django-cors-headers
drf-haystack
elasticsearch==2.4.1